refator: extracted title of pages as reusable and consistent component
