+++ frontend/src/components/PageHeader.jsx
@@ -1,0 +1,14 @@
+import {Box, Text} from "@chakra-ui/react";
+import React from "react";
+
+const PageHeader = ({fontSize, pageDesc}) => {
+    return (
+        <Box as="header" p="4" textAlign="center">
+            <Text fontSize={fontSize} fontWeight="bold">
+                {pageDesc}
+            </Text>
+        </Box>
+    );
+}
+
+export default PageHeader;
