+++ frontend/src/pages/GroupAnalytics.jsx
@@ -18,6 +18,7 @@ import { Chart as ChartJS, ArcElement, Tooltip, Legend, LinearScale, CategorySca
 import { ArrowBackIcon } from '@chakra-ui/icons';
 import { MockAuth } from '../mockAuth/mockAuth';
 import { useAuth0 } from '@auth0/auth0-react';
+import PageHeader from "../components/PageHeader";
 
 ChartJS.register(ArcElement, Tooltip, Legend, LinearScale, CategoryScale, BarController, BarElement);
 
@@ -70,11 +71,10 @@ const UnitAnalytics = () => {
 
     return (
         <div>
-            <Box as="header" p="4" textAlign="center">
-                <Text fontSize="2xl" fontWeight="bold">
-                    {`Configure Groups for: ${unitCode} - ${period}, ${year}`}
-                </Text>
-            </Box>
+            <PageHeader
+                fontSize={"2xl"}
+                pageDesc={`Group analytics: ${unitCode} ${period} ${year}`}
+            />
 
             <Center>
                 <Button onClick={navigateToOfferingDashboard}>
