+++ frontend/src/pages/Students.jsx
@@ -163,6 +163,7 @@ function Students() {
             <Heading alignContent={'center'}>
                 <Center margin="10">{`${unitCode} - ${period}, ${year}`}</Center>
             </Heading>
+
             <Button
                 me="12px"
                 align="right"
